Labo3 JS

+++ JS/Labo3JS/Opdracht Producten/scripts/code.js
@@ -18,9 +18,9 @@ const update =() => {
     for (let i = 0; i < txtSubtotaal.length; i++) {
 
 
-        let prijs = parseFloat(txtPrijs[i].innerHTML);
-        let aantal = parseFloat(txtAantal[i].value);
-        let btw = parseFloat(txtBtw[i].innerHTML);
+        let prijs = parseInt(txtPrijs[i].innerHTML);
+        let aantal = parseInt(txtAantal[i].value);
+        let btw = parseInt(txtBtw[i].innerHTML);
         let subtotaal = aantal * prijs * (1 + (btw / 100));
         totaal = subtotaal + totaal;
